+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.0.4
+* Permission and full path fixes update fullpath method & add permission for example app #1 by https://github.com/lawnvi and implement PWD 
+* Added logger method to allow custom logs handling
+* Added server startInBackground method
+* General cleanup and adding test cases
 
+## 0.0.3
+* Refactored to allow custom logs 
 ## 0.0.2
 * Added CDUP command 
 ## 0.0.1+1